Change test style to AnyWordSpec
Also, implement tests for the simplest case (simple parser)
